% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceNetwork.R
\name{forceNetwork}
\alias{forceNetwork}
\title{<Add Title>}
\usage{
forceNetwork(
  data = jsonlite::parse_json(paste(readLines(sys_file("extdata/forceData.json")),
    collapse = "")),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
